requirements.txt
asgiref==3.4.1
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
charset-normalizer==2.0.7
distlib==0.3.3
dj-database-url==0.5.0
Django==3.2.8
django-cors-headers==3.10.0
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
filelock==3.3.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.1
Pillow==8.4.0
platformdirs==2.4.0
psycopg2==2.9.1
pycodestyle==2.8.0
PyJWT==2.3.0
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
virtualenv==20.8.1
whitenoise==5.3.0
zipp==3.6.0
